# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "FISBook Reviews API",
    "description": "Welcome to the API documentation for the FISBook Reviews API. This API allows you to interact with the reviews of books and reading lists in the FISBook application."
  },
  "paths": {
    "/": {
      "get": {
        "description": "INDEX.js endpoint",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/reviews/books": {
      "get": {
        "summary": "Get all the reviews of every book",
        "description": "Get all the reviews of books in the database",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a new review.",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "example": "1"
                  },
                  "book_id": {
                    "example": "1"
                  },
                  "score": {
                    "example": "5"
                  },
                  "title": {
                    "example": "Nice book"
                  },
                  "comment": {
                    "example": "I really enjoyed this book"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/reading_lists": {
      "get": {
        "description": "Get all the reviews of every reading list",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "Create a review for a reading list",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "example": "1"
                  },
                  "reading_list_id": {
                    "example": "1"
                  },
                  "score": {
                    "example": "3"
                  },
                  "comment": {
                    "example": "Not bad!"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/books/bk/{bookID}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "bookID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/reviews/reading_lists/rl/{readingListID}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "readingListID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/reviews/reading_lists/rev/{reviewID}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "reviewID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/reviews/books/rev/{reviewID}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "reviewID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/reviews/books/{reviewID}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "reviewID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "score": {
                    "example": "any"
                  },
                  "title": {
                    "example": "any"
                  },
                  "comment": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "reviewID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/reviews/reading_lists/{reviewID}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "reviewID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "score": {
                    "example": "any"
                  },
                  "comment": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "reviewID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}